// packages/core-react-theme/src/types/colors.ts
export interface IThemeColors {
  background: {
    note: string
    info: string
    success: string
    warning: string
    error: string

    body: string

    blockquote: string
    code: string
    codeHighlight: string
    inlineCode: string
    tableHead: string
    tableOddRow: string
    tableEvenRow: string
  }
  border: {
    note: string
    info: string
    success: string
    warning: string
    error: string

    blockquote: string
    code: string
    codeLineno: string
    heading: string
    image: string
    table: string
    thematicBreak: string
  }
  text: {
    note: string
    info: string
    success: string
    warning: string
    error: string

    body: string

    codeCaret: string
    codeTitle: string
    codeSelection: string
    delete: string
    heading: string
    imageTitle: string
    inlineCode: string
    link: string
    math: string
  }
}